Возвращает компонент карты, удаляет hoc
